requirements.txt
apache-airflow-providers-google==17.1.0
apache-airflow-providers-slack==9.1.4
google-cloud-storage<=3.0.0
